refactor: rpc namespace refactor - personal
